ch(fastfoodfast): organised files into modules
Restructure app directory to improve readability

[Delivers #160982299]
